Fix state updates triggered by actions
The state objects in `STATES` were overwritten with a new object
whenever a state was received from the broker.  This lead to
inconsistencies as the objects in `STATES` were no longer the same as
those captured in the handlers in `STATE_ACTION_HANDLERS`.  This
resulted in the processes not seeing the updated objects.

This change eliminates the creation of a new element by writing the
fields of the state individually.